+++ src/main/java/com/tcs/abn/recipeapi/entity/RecRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RecRepository extends JpaRepository<Recipe, Integer>{
+public interface RecRepository extends JpaRepository<Recipe, Long>{
 
 }
